requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
Django==4.2.4
django-extensions==3.2.3
django-guardian==2.4.0
django-redis==5.3.0
django-reversion==5.0.4
djangorestframework==3.14.0
exceptiongroup==1.1.2
factory-boy==3.3.0
Faker==19.3.0
iniconfig==2.0.0
packaging==23.1
Pillow==10.0.0
pluggy==1.2.0
psycopg2-binary==2.9.7
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.7.1
